Refactor source error states into new component
